Modernize wrapper for BaseRecord/BaseCatalog.
