+++ python/lsst/afw/table/_base.py
@@ -21,8 +21,8 @@
 import numpy as np
 
 from lsst.utils import continueClass, TemplateMeta
-from .base import BaseRecord, BaseCatalog
-from .._schema import Key
+from ._table import BaseRecord, BaseCatalog
+from ._schema import Key
 
 
 __all__ = ["Catalog"]
