+++ python/lsst/afw/table/catalogMatches.py
@@ -29,7 +29,7 @@ import numpy as np
 import lsst.pex.exceptions as pexExcept
 from ._schema import Schema
 from ._schemaMapper import SchemaMapper
-from .base import BaseCatalog
+from ._base import BaseCatalog
 from .simple import SimpleCatalog, SimpleTable
 from .source import SourceCatalog, SourceTable
 from .match import ReferenceMatch
