+++ python/lsst/afw/table/match/matchContinued.py
@@ -23,7 +23,7 @@ __all__ = ["SimpleMatch", "ReferenceMatch", "SourceMatch", "clone", "matchCls", 
 
 import numpy as np
 
-from ..base import BaseCatalog
+from .._base import BaseCatalog
 from .._schema import Schema
 from .match import SimpleMatch, ReferenceMatch, SourceMatch
